requirements.dev.txt
-r requirements.txt

pytest==3.6.2
pytest-cov==2.5.1
pytest-django==3.3.2
